@@ -3,10 +3,10 @@ package com.kyc.ondato.ondato_flutter
 import android.content.Context
 import android.util.Log
 import androidx.annotation.NonNull
-import com.kyc.ondato.Ondato
-import com.kyc.ondato.OndatoConfig
-import com.kyc.ondato.OndatoError
-import com.kyc.ondato.enums.Language
+import com.ondato.sdk.Ondato
+import com.ondato.sdk.OndatoConfig
+import com.ondato.sdk.OndatoError
+import com.ondato.sdk.enums.Language
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodCall
@@ -53,12 +53,15 @@ class OndatoFlutterPlugin: FlutterPlugin, MethodCallHandler {
               }
 
               override fun onFailure(identificationId: String?, error: OndatoError) {
-                  result.success(
-                      mapOf(
-                          "identificationId" to identificationId,
-                          "error" to error.message
-                      )
-                  )
+                 val message = error.message;
+                 Log.d("ONDATO PLUGIN", "ERROR FULL $error")
+                 Log.d("ONDATO PLUGIN", "ERROR MESSAGE $message")
+                   result.success(
+                       mapOf(
+                           "identificationId" to identificationId,
+                           "error" to error.message
+                       )
+                   )
               }
           })
       } catch (e: Exception) {
@@ -70,37 +73,23 @@ class OndatoFlutterPlugin: FlutterPlugin, MethodCallHandler {
     val map = args as? HashMap<String, Any>
     if (map != null) {
         val mode = map["mode"] as String
-        val credentials = map["credentials"] as HashMap<String?, String?>
+        val identificationId = map["identificationId"] as String
         val flowConfig = map["flowConfiguration"] as? HashMap<String, Boolean>
         val language = map["language"] as String
 
-        val identificationId = credentials["identificationId"]?: ""
-        Log.d("ONDATO SDK", "identificationID $identificationId")
+        Log.d("ONDATO PLUGIN", "identificationID $identificationId")
 
         val config = OndatoConfig.Builder()
-                .recordProcess(flowConfig?.get("recordProcess") ?: false) //defaul is false
-                .setToken(credentials["accessToken"] ?: "")
-                .setIdentificationId(identificationId)
-                .showSplashScreen(flowConfig?.get("showSplashScreen") ?: true) //default is true
-                .showStartScreen(flowConfig?.get("showStartScreen") ?: true) //default is true
-                .showConsentScreen(flowConfig?.get("showConsentScreen")
-                        ?: true) //default is true
-                .showSelfieWithDocumentScreen(flowConfig?.get("showSelfieAndDocumentScreen")
-                        ?: true) //default is true
-                .showSuccessScreen(flowConfig?.get("showSuccessWindow")
-                        ?: true) //default is true
-                .ignoreLivenessError(flowConfig?.get("ignoreLivenessErrors")
-                        ?: false) //
-                .ignoreVerificationErrors(flowConfig?.get("ignoreVerificationErrors")
-                        ?: false) //default is false
-                .recordProcess(flowConfig?.get("recordProcess") ?: false) //default is true
-                .setMode(getMode(mode)) //default is TEST
-                .setLanguage(getLanguage(language)) // default is English
-                .build()
+            .setIdentityVerificationId(identificationId) //should by generated by user and passed here
+            .showStartScreen(flowConfig?.get("showSplashScreen") ?: true) //default is true
+            .showSplashScreen(flowConfig?.get("showStartScreen") ?: true) //default is true
+            .setMode(getMode(mode)) //default is TEST
+            .setLanguage(getLanguage(language))
+            .build()
         Ondato.init(config)
         result.success(true)
     } else {
-        result.error(null, "", "Missing parameters")
+        result.error("", "", "Missing parameters")
     }
   }
 
