@@ -1,4 +1,7 @@
+// ignore_for_file: non_constant_identifier_names
+
 import 'dart:core';
+import 'dart:developer';
 
 import 'package:flutter/material.dart';
 import 'package:ondato_flutter/ondato_config.dart';
@@ -9,44 +12,47 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
+  String get YOUR_IDENTIFICATION_ID => 'YOUR_IDENTIFICATION_ID';
+
+  Future<void> startIdentification() async {
+    try {
+      await OndatoFlutter.init(
+        OndatoServiceConfiguration(
+          identificationId: YOUR_IDENTIFICATION_ID,
+          language: OndatoLanguage.en,
+          mode: OndatoEnvironment.test,
+          flowConfiguration: OndatoFlowConfiguration(
+            showSplashScreen: true,
+            showStartScreen: true,
+          ),
+          appearance: OndatoIosAppearance(
+            errorColor: Colors.orange,
+            progressColor: Colors.orange,
+          ),
+        ),
+      );
+      var identificationId = await OndatoFlutter.startIdentification();
+      log('Success with ${identificationId!} IdentificationId');
+    } catch (e) {
+      if (e is OndatoException) {
+        log('Exception ${e.error.toString()}');
+      } else {
+        log(e.toString());
+      }
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
-    const YOUR_ACCESS_TOKEN = '';
-    const YOUR_IDENTIFICATION_ID = '';
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Ondato Fluttter example app'),
+          title: const Text('Ondato Flutter example app'),
         ),
         body: Center(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              TextButton(
-                onPressed: () async {
-                  await OndatoFlutter.init(
-                    OndatoServiceConfiguration(
-                      appearance: OndatoIosAppearance(
-                        errorColor: Colors.orange,
-                        progressColor: Colors.orange,
-                      ),
-                      language: OndatoLanguage.en,
-                      credentials: OndataCredencials(
-                          accessToken: YOUR_ACCESS_TOKEN,
-                          identificationId: YOUR_IDENTIFICATION_ID),
-                      flowConfiguration: OndatoFlowConfiguration(
-                          recordProcess: false,
-                          ignoreLivenessErrors: true,
-                          ignoreVerificationErrors: true),
-                      mode: OndatoEnvironment.test,
-                    ),
-                  );
-                  await OndatoFlutter.startIdentification()
-                      .then((indetificationId) => print(indetificationId));
-                },
-                child: Text('Start Identification'),
-              ),
-            ],
+          child: TextButton(
+            onPressed: () => startIdentification(),
+            child: Text('Start Identification'),
           ),
         ),
       ),
